+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v2.0.0
-
+- Removed TTK_AssertR and TTK_AssertMR.
 # v1.0.2 (26-09-2022)
 - Updated documentation.
 # v1.0.1 (25-09-2022)